requirements.txt
apache-airflow-providers-cncf-kubernetes==7.4.2
apache-airflow==2.6.2
pyspark==3.4.0